requirements.txt
streamlit
openai
groq
python-dotenv
SpeechRecognition
pyttsx3
PyAudio
PyQt6
